+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - getCloudinaryUrl service function
 - Methods for a few image transformations like pad, limit, scale ...
 - A few effect methods like grayscale, blur, sepia ...
+- Readonly behaviour
 
 ## Fixed
 - Use the service function to get the cloudinary url within the CloudinaryImage Link() function (ensures proper credentials setup)